Remove test scanner from the ScanningStore class
- Remove the scanner definition from the class' init method
- Add the scanner in the test fixtures where necessary
- Move the tests of the scanjobs API to integration
