+++ tests/integration/api/conftest.py
@@ -6,10 +6,11 @@ import pytest
 from flask import Flask
 
 from scanomatic.io.imagestore import ImageStore
-from scanomatic.ui_server import scanners_api
+from scanomatic.io.scanning_store import ScanningStore
+from scanomatic.models.scanner import Scanner
 from scanomatic.ui_server import scan_jobs_api
+from scanomatic.ui_server import scanners_api
 from scanomatic.ui_server import scans_api
-from scanomatic.ui_server.ui_server import add_configs
 
 
 @pytest.fixture
@@ -18,7 +19,13 @@ def app(tmpdir):
     app.register_blueprint(scanners_api.blueprint, url_prefix="/scanners")
     app.register_blueprint(scan_jobs_api.blueprint, url_prefix="/scan-jobs")
     app.register_blueprint(scans_api.blueprint, url_prefix="/scans")
-    add_configs(app)
+    app.config['scanning_store'] = ScanningStore()
+    app.config['scanning_store'].add_scanner(
+        Scanner('Scanner one', '9a8486a6f9cb11e7ac660050b68338ac')
+    )
+    app.config['scanning_store'].add_scanner(
+        Scanner('Scanner two', '350986224086888954')
+    )
     app.config['imagestore'] = ImageStore(str(tmpdir))
     return app
